Move shuffle to separate helper file
