+++ app/controllers/quiz/question.js
@@ -1,6 +1,7 @@
 import Controller from '@ember/controller';
 import { inject as service } from '@ember/service';
 import { computed } from '@ember/object';
+import shuffle from '../../helpers/shuffle';
 
 export default Controller.extend({
   cookies: service(),
@@ -47,23 +48,3 @@ export default Controller.extend({
     }
   }
 });
-
-function shuffle(array) {
-  let counter = array.length;
-
-  // While there are elements in the array
-  while (counter > 0) {
-      // Pick a random index
-      let index = Math.floor(Math.random() * counter);
-
-      // Decrease counter by 1
-      counter--;
-
-      // And swap the last element with it
-      let temp = array[counter];
-      array[counter] = array[index];
-      array[index] = temp;
-  }
-
-  return array;
-}
